@@ -7,7 +7,7 @@ class LogcatMonitor {
       const MethodChannel('logcat_monitor/methods');
   static const EventChannel _stream =
       const EventChannel('logcat_monitor/events');
-  static StreamSubscription _streamSubscription;
+  static late StreamSubscription _streamSubscription;
 
   static void addListen(void Function(dynamic) onData) {
     _streamSubscription = _stream.receiveBroadcastStream().listen(onData);
@@ -17,25 +17,25 @@ class LogcatMonitor {
     _streamSubscription.cancel();
   }
 
-  static Future<bool> startMonitor(String options) async {
-    bool result = await _channel
+  static Future<bool?> startMonitor(String options) async {
+    bool? result = await _channel
         .invokeMethod('startMonitor', <String, String>{'options': options});
     return result;
   }
 
-  static Future<bool> stopMonitor() async {
-    bool result = await _channel.invokeMethod('stopMonitor');
+  static Future<bool?> stopMonitor() async {
+    bool? result = await _channel.invokeMethod('stopMonitor');
     return result;
   }
 
-  static Future<String> get getLogcatDump async {
-    final String result = await _channel
+  static Future<String?> get getLogcatDump async {
+    final String? result = await _channel
         .invokeMethod('runLogcat', <String, String>{'options': "-d"});
     return result;
   }
 
-  static Future<String> get clearLogcat async {
-    final String result = await _channel
+  static Future<String?> get clearLogcat async {
+    final String? result = await _channel
         .invokeMethod('runLogcat', <String, String>{'options': "-c"});
     return result;
   }
