@@ -3,13 +3,25 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:logcat_monitor/logcat_monitor.dart';
 
 void main() {
-  const MethodChannel channel = MethodChannel('logcat_monitor');
+  const MethodChannel channel = MethodChannel('logcat_monitor/methods');
 
   TestWidgetsFlutterBinding.ensureInitialized();
 
   setUp(() {
     channel.setMockMethodCallHandler((MethodCall methodCall) async {
-      return '42';
+      switch (methodCall.method) {
+        case 'startMonitor':
+          return true;
+          break;
+        case 'stopMonitor':
+          return true;
+          break;
+        case 'runLogcat':
+          return 'logcat example';
+          break;
+        default:
+          return 'not implemented';
+      }
     });
   });
 
@@ -17,7 +29,7 @@ void main() {
     channel.setMockMethodCallHandler(null);
   });
 
-  test('getPlatformVersion', () async {
-    expect(await LogcatMonitor.platformVersion, '42');
+  test('getLogcatDump', () async {
+    expect(await LogcatMonitor.getLogcatDump, 'logcat example');
   });
 }
